--- src/common/client_test.go
@@ -1,1 +1,0 @@
-package common_test
